adc works. adch and adcl addressed dont work cause fucking arduino

+++ module4/include/AnalogPin.h
@@ -31,17 +31,17 @@ class AnalogPin
 {
     public:
         int _adc;
-        int _count;
+        uint16_t _count;
 
         AnalogPin(analogpin_number pin);
         //AnalogPin(port_t port, int pin);
-        int read();
+        uint16_t read();
         int select_ADC();
         int set_ref(uint8_t ref);
         int set_ADC(bool state);
         int set_prescaler(uint8_t ps);
         int start_conv();
-        int get_count();
+        uint16_t get_count();
         void print();
     private:
         bool _valid;
